refactor: fix links
