--- src/modules/auth/__mocks__/index.ts
@@ -1,3 +1,0 @@
-export * from "./auth.service";
-export * from "./config-service.mock";
-export * from "./jwt-service.mock";
