--- src/modules/locations/__mocks__/index.ts
@@ -1,1 +1,0 @@
-export * from "./prisma-service.mock";
