--- src/modules/pets/__mocks__/index.ts
@@ -1,2 +1,0 @@
-export * from "./pets.service";
-export * from "./prisma-service.mock";
