--- src/modules/species/__mocks__/index.ts
@@ -1,2 +1,0 @@
-export * from "./species.service";
-export * from "./prisma-service.mock";
