--- src/modules/users/__mocks__/index.ts
@@ -1,2 +1,0 @@
-export * from "./users.service";
-export * from "./prisma-service.mock";
